Collect statistics form `perf`
This commit implements collection of statistics and hardware counters through
`perf`, available on Linux. This adds a new field to the `Run` struct which
may have a list of stats, and we can aggregate these stats over time.

A new "fake rustc" shim was added which is run in lieu of the actual rustc. When
the shim detects `time-passes` is passed it assumes this is the "relevant rustc"
to profile, so it executes the compiler inside of `perf stat` to collect
statistics.

Currently this counts user-level instructions and cycles, wall clocks, and
faults. The thinking is that the instructions/cycles may be pretty smooth over
time, the wall clock provides a second source of time to the summation of `-Z
time-passes`, and the faults may let us know when memory is balooning.

As a side implementation detail this also implements the automatically running
the final compiler as high priority and pinning it to one CPU, again in the
hopes of leveling out the counters.